% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parental_clade.r
\name{check_parent_clade}
\alias{check_parent_clade}
\title{Check if parental clade of node is fully contained in ids}
\usage{
check_parent_clade(tre, node, ids)
}
\arguments{
\item{tre}{}

\item{node}{}

\item{ids}{}
}
\value{

}
\description{
Internal
}
