% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mono_clades.r
\name{find_mono_clades}
\alias{find_mono_clades}
\title{Find monophyletic clades}
\usage{
find_mono_clades(tre, ids, return = "all")
}
\arguments{
\item{tre}{A phylo object}

\item{ids}{A vector of tip labels or node IDs}

\item{return}{Either "all" nodes or "outgroup" and tips.}
}
\value{
A list of numeric vectors where each vector is the
node IDs that correspond to a clade in the tree where all
the tips are contained in ids.
}
\description{
Takes a phylo object and a vector of tip labels. Identifies
all the clades in the phylogeny that have all their tips
contained in the given list.
}
