% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_children.r
\name{node_children}
\alias{node_children}
\title{Node childre from phylo object}
\usage{
node_children(tre, node)
}
\arguments{
\item{tre}{A phylo object}

\item{node}{Numeric ID or tip label}
}
\value{
Numeric vector of IDs
}
\description{
Node childre from phylo object
}
