% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_descendants.r
\name{node_descendants}
\alias{node_descendants}
\title{All descendants from node in phylo object}
\usage{
node_descendants(tre, node)
}
\arguments{
\item{tre}{Phylo object}

\item{node}{Node ID or tip label.}
}
\value{

}
\description{
All descendants from node in phylo object
}
