% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_parent.r
\name{node_parent}
\alias{node_parent}
\title{Node parent in phylo object}
\usage{
node_parent(tre, node)
}
\arguments{
\item{tre}{Phylo object}

\item{node}{Numeric ID or tip label}
}
\value{
Numeric node ID.
}
\description{
Node parent in phylo object
}
