% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parental_clade.r
\name{recursive_check_parent_clade}
\alias{recursive_check_parent_clade}
\title{Recursive check until clade is not in ids}
\usage{
recursive_check_parent_clade(tre, node, ids)
}
\arguments{
\item{tre}{}

\item{node}{}

\item{ids}{}
}
\value{

}
\description{
Internal
}
